lordran/requirements.txt
djangorestframework
pillow
drf_yasg
django-cors-headers
djangorestframework-simplejwt
